requirements.txt
PySide6==6.8.3
PyOpenGL==3.1.10
Jinja2==3.1.6
PyYAML==6.0.3